Update drawpoints.cpp
I have added the option to report networks to a file in a "condensed" format. In this format, the first line is a manifest giving the number of points, the number of edges,
and a number which essentially gives the width of the network. This is used to express the periodicity in the horizontal direction for networks with bonds that connect a vertex on the left edge to a vertex on the right edge. After the initial manifest, the file specifies each point on its own line as an x-y coordinate pair. Finally, each edge is specified as the integer indices into the list of points of its end points, followed by a value that indicates if the edge does not wrap around (0) wraps from left to right (-1), or wraps from right to left (1).